LeetCode 237: Delete Node
According to LeetCode, you don’t need to travel through the LinkedList
and find out which node to delete. Instead, it just assume you are
already at the Node that you need to delete and ask you do the work of
deleting it. (i.e. Handling the next pointer and value change.)